feat: separated Hero section home in a independent component for WIP modifications.
