+++ front/reyesandfriends.cl/src/pages/home/home.tsx
@@ -1,44 +1,13 @@
-import React, { useEffect, useState } from "react";
-import { ChevronRight, } from "lucide-react";
+import React from "react";
+import { ChevronRight } from "lucide-react";
 import { Link } from "react-router-dom";
-import { AnimatePresence, motion } from "framer-motion";
 import TerminalMessage from "./components/TerminalMessage";
 import MethodologySteps from "./components/MethodologySteps";
 import WhyChooseReyesAndFriends from "./components/WhyChooseReyesAndFriends";
-import AnimatedMockups from "./components/AnimatedMockups";
+import HomeHeroSection from "./components/HomeHeroSetion";
 import { Helmet } from "react-helmet-async";
 
 const Home: React.FC = () => {
-
-    const subtitles = [
-        "tu crecimiento digital",
-        "tu éxito en línea",
-        "tu presencia web",
-        "tu transformación digital",
-        "tu innovación tecnológica",
-        "tu futuro digital",
-        "tu negocio en la nube",
-        "tu solución tecnológica",
-        "tu plataforma digital",
-        "tu estrategia en línea",
-        "tu imperio digital",
-        "tu marca en línea",
-        "tu proyecto web",
-        "tu sueño digital",
-    ];
-
-    const [currentSubtitleIndex, setCurrentSubtitleIndex] = useState(0);
-
-    useEffect(() => {
-        const interval = setInterval(() => {
-            setCurrentSubtitleIndex((prevIndex) => 
-                (prevIndex + 1) % subtitles.length
-            );
-        }, 4000);
-
-        return () => clearInterval(interval);
-    }, [subtitles.length]);
-    
     return (
         <>
             <Helmet>
@@ -57,60 +26,7 @@ const Home: React.FC = () => {
                 <meta name="twitter:image" content="/img/open-graph-images/default-style.png" />
             </Helmet>
             
-            <section className="bg-cover bg-center relative min-h-[800px] md:min-h-[800px] flex items-center">
-                <div className="absolute inset-0 bg-hero-section z-10"></div>
-                <div className="absolute inset-0">
-                    <img
-                        src="/img/background/background-web.jpg"
-                        alt="Ilustración de fondo - Edificios"  
-                        className="w-full h-full object-cover filter grayscale z-0"
-                        draggable={false}
-                        onContextMenu={e => e.preventDefault()}
-                    />
-                </div>
-                <div className="container mx-auto px-4 py-24 relative z-10 flex flex-col md:flex-row items-center justify-center text-center md:text-left flex-1">
-                    <div className="flex-1 flex flex-col items-center md:items-start justify-center max-w-2xl mx-auto">
-                        <h1 className="text-5xl md:text-6xl mb-2 text-white drop-shadow-lg tracking-tight">
-                            Impulsamos
-                        </h1>
-                        <h1 className="text-4xl md:text-5xl mb-6 inline-block text-white drop-shadow">
-                            <AnimatePresence mode="wait" initial={false}>
-                                <motion.span
-                                    key={subtitles[currentSubtitleIndex]}
-                                    initial={{ opacity: 0, y: 20 }}
-                                    animate={{ opacity: 1, y: 0 }}
-                                    exit={{ opacity: 0, y: -20 }}
-                                    transition={{ 
-                                        duration: 0.3,
-                                        ease: "easeInOut"
-                                    }}
-                                    className="text-red-600"
-                                >
-                                    {subtitles[currentSubtitleIndex]}
-                                </motion.span>
-                            </AnimatePresence>
-                        </h1>
-                        <div className="mb-8 w-full flex flex-col items-center">
-                            <p className="text-2xl text-white drop-shadow-sm">
-                                ¡Transforma tu negocio con tecnología a tu medida!<br />
-                                Creamos sitios web, plataformas y soluciones digitales innovadoras para que tu empresa crezca y destaque en el mundo digital.
-                            </p>
-                        </div>
-                        <div className="flex flex-col gap-4 justify-center md:justify-start">
-                            <Link
-                                to="/quote-project"
-                                className="bg-red-700 hover:bg-red-800 text-white px-10 py-5 rounded transition-all duration-300 flex items-center gap-4 font-semibold hover:scale-105 text-2xl"
-                            >
-                                Trabajemos juntos
-                                <ChevronRight className="h-7 w-7" />
-                            </Link>
-                        </div>
-                    </div>
-                    <div className="hidden md:flex flex-1 justify-center items-center ml-8">
-                        <AnimatedMockups />
-                    </div>
-                </div>
-            </section>
+            <HomeHeroSection />
 
             <section className="py-16 bg-zinc-900 text-white">
                 <div className="container mx-auto px-4 max-w-7xl">
@@ -164,7 +80,6 @@ const Home: React.FC = () => {
             </section>
 
             <WhyChooseReyesAndFriends />
-
         </>
     );
 };
